+++ gameplay/src/Slider.h
@@ -207,7 +207,7 @@ protected:
     /**
      * @see Control::MouseEvent
      */
-    bool mouseEvent(Mouse::MouseEvent evt, int x, int y, int wheelDelta);
+    bool mouseEvent(Mouse::MouseEvent evt, int x, int y, float wheelDelta);
 
     /**
      * @see Control::gamepadJoystickEvent
